+++ xmhf/src/libbaremetal/libtv_utpm/include/tv_utpm.h
@@ -268,7 +268,7 @@ TPM_RESULT utpm_unseal(utpm_master_state_t *utpm, uint8_t* input, uint32_t inlen
 
 TPM_RESULT utpm_quote(TPM_NONCE* externalnonce, TPM_PCR_SELECTION* tpmsel, /* hypercall inputs */
                       uint8_t* output, uint32_t* outlen, /* hypercall outputs */
-                      uint8_t* pcrComp, uint32_t* pcrCompLen,                      
+                      uint8_t* pcrComp, uintptr_t* pcrCompLen,                      
                       utpm_master_state_t *utpm); /* TrustVisor inputs */
 
 /**
